+++ test/rule_test.dart
@@ -1,28 +1,64 @@
 // ignore_for_file: non_constant_identifier_names
 
+import 'dart:io';
+
+import 'package:analyzer/src/lint/registry.dart';
 import 'package:analyzer_testing/analysis_rule/analysis_rule.dart';
+import 'package:prefer_shorthands/main.dart';
+import 'package:prefer_shorthands/settings.dart';
 import 'package:test_reflective_loader/test_reflective_loader.dart';
 
-void main() {
+late final String codeContent;
+
+void main() async {
+  Registry.ruleRegistry.registerLintRule(PreferShorthandsRule());
+
+  codeContent = await File('example/lib/main.dart').readAsString();
+
   defineReflectiveSuite(() {
     defineReflectiveTests(PreferShorthandsRuleTest);
   });
 }
 
-// TODO: test not working yet, because i can't find a API to register the rule
 @reflectiveTest
 class PreferShorthandsRuleTest extends AnalysisRuleTest {
   @override
   String get analysisRule => 'prefer_shorthands';
 
   void test_prefer_shorthands() async {
-    await assertDiagnostics(
-      r'''
-final SomeEnum temp = SomeEnum.a;
-''',
-      [lint(33, 5)],
-    );
+    plugin.settings = Settings(convertImplicitDeclaration: false);
+    await assertDiagnostics(codeContent, [
+      lint(200, 11),
+      lint(353, 7),
+      lint(372, 7),
+      lint(400, 5),
+      lint(410, 11),
+      lint(439, 5),
+      lint(446, 11),
+      lint(497, 5),
+      lint(512, 5),
+    ]);
+  }
+
+  void test_convert_implicit_declaration() async {
+    plugin.settings = Settings(convertImplicitDeclaration: true);
+    await assertDiagnostics(codeContent, [
+      lint(32, 23),
+      lint(86, 10),
+      lint(111, 5),
+      lint(131, 5),
+      lint(151, 7),
+      lint(173, 10),
+      lint(200, 11),
+      lint(353, 7),
+      lint(372, 7),
+      lint(400, 5),
+      lint(410, 11),
+      lint(439, 5),
+      lint(446, 11),
+      lint(473, 30),
+      lint(497, 5),
+      lint(512, 5),
+    ]);
   }
 }
-
-enum SomeEnum { a, b }